fashion-classifier/app/requirements.txt
pillow==11.1.0
requests==2.32.3
numpy==1.26.4
tflite-runtime==2.7.0